Add an explicit cast to avoid a compiler warning
readerfactory.c:1033:45: warning: comparison of integers of different signs:
      'unsigned long' and 'SCARDHANDLE' (aka 'long') [-Wsign-compare]
                                if ((rContext->dwIdentity + randHandle) ==
                                    ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ ^


git-svn-id: svn://anonscm.debian.org/svn/pcsclite/trunk/PCSC@5517 0ce88b0d-b2fd-0310-8134-9614164e65ea

+++ src/readerfactory.c
@@ -1030,7 +1030,7 @@ again:
 			while (list_iterator_hasnext(l))
 			{
 				currentHandle = list_iterator_next(l);
-				if ((rContext->dwIdentity + randHandle) ==
+				if (((LONG)rContext->dwIdentity + randHandle) ==
 					(currentHandle->hCard))
 				{
 					/* Get a new handle and loop again */
